# kf01/api-contracts/openapi.yaml
openapi: 3.1.0
info:
  title: 往日重现 2.0 API Contract
  version: 0.2.0
  description: >
    2.0 契约草案。当前 NestJS 后端先按该契约演进，后续 FastAPI + PostgreSQL
    迁移时保持字段兼容。
servers:
  - url: http://localhost:3999/api
    description: local NestJS bridge
paths:
  /auth/login:
    post:
      summary: TV 端账号密码登录往日重现后端
      tags: [TV Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: 登录成功并返回设备令牌
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
...
  /admin/photo-library/playback-albums:
    get:
      summary: 管控端播放相册列表
      tags: [Admin Playback Albums]
      responses:
        '200':
          description: 播放相册列表
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlaybackAlbumListEnvelope'
    post:
      summary: 新建播放相册
      tags: [Admin Playback Albums]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePlaybackAlbumRequest'
      responses:
        '200':
          description: 创建结果
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlaybackAlbumEnvelope'
  /admin/photo-library/playback-albums/{playbackAlbumId}/photos:
    get:
      summary: 查看播放相册照片，包含本地成员和按需读取的飞牛相册照片
      tags: [Admin Playback Albums]
      parameters:
        - $ref: '#/components/parameters/PlaybackAlbumId'
      responses:
        '200':
          description: 播放相册成员
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PhotoCenterItemsEnvelope'
  /admin/photo-library/playback-albums/{playbackAlbumId}/ai-policy:
    patch:
      summary: 更新相册 AI 开关和策略，开启时触发按需导入与 AI 持久化
      tags: [Admin AI]
      parameters:
        - $ref: '#/components/parameters/PlaybackAlbumId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePlaybackAlbumAiPolicyRequest'
      responses:
        '200':
          description: AI 策略更新结果
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlaybackAlbumAiPolicyUpdateEnvelope'
  /admin/photo-library/playback-albums/{playbackAlbumId}/ai-jobs:
    post:
      summary: 手动触发播放相册 AI 任务
      tags: [Admin AI]
      parameters:
        - $ref: '#/components/parameters/PlaybackAlbumId'
      responses:
        '200':
          description: AI 任务结果
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlaybackAlbumAiJobEnvelope'
  /admin/photo-library/photos/{photoId}/ai-sync:
    post:
      summary: 将照片已保存的 AI 原始返回重新解析并同步到结构化字段，不调用模型
      tags: [Admin AI]
      parameters:
        - name: photoId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 同步后的照片详情
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PhotoCenterItemEnvelope'
  /admin/photo-library/ai-settings:
    get:
      summary: 获取 AI 设置中心配置
      tags: [Admin AI]
      responses:
        '200':
          description: AI 设置
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AiSettingsEnvelope'
    put:
      summary: 保存 AI 平台、模型、接口地址、API Key 和提示词
      tags: [Admin AI]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAiSettingsRequest'
      responses:
        '200':
          description: AI 设置
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AiSettingsEnvelope'
...
components:
  securitySchemes:
    deviceToken:
      type: apiKey
      in: header
      name: x-device-token
  parameters:
    PlaybackAlbumId:
      name: playbackAlbumId
      in: path
      required: true
      schema:
        type: string
  schemas:
    EnvelopeBase:
      type: object
      properties:
        code:
          type: integer
          const: 0
      required: [code]
    LoginRequest:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required: [username, password]
    LoginResponse:
      allOf:
        - $ref: '#/components/schemas/EnvelopeBase'
        - type: object
          properties:
            data:
              type: object
              properties:
                deviceToken:
                  type: string
                expiresAt:
                  type: string
                  format: date-time
                username:
                  type: string
              required: [deviceToken]
          required: [data]
...
    PlaybackAlbum:
      type: object
      properties:
        playbackAlbumId:
          type: string
        title:
          type: string
        sourceType:
          type: string
          enum: [manual, feiniu_album]
        sourceAlbumId:
          type: string
        sourceAlbumTitle:
          type: string
        aiEnabled:
          type: boolean
        aiScoreThreshold:
          type: integer
        aiPriorityTags:
          type: array
          items:
            type: string
        pushEnabled:
          type: boolean
        pushScoreThreshold:
          type: integer
        pushPriorityTags:
          type: array
          items:
            type: string
        photoCount:
          type: integer
      required: [playbackAlbumId, title, sourceType, aiEnabled, photoCount]
    CreatePlaybackAlbumRequest:
      type: object
      properties:
        title:
          type: string
        sourceType:
          type: string
          enum: [manual, feiniu_album]
        sourceAlbumId:
          type: string
        sourceAlbumTitle:
          type: string
        aiEnabled:
          type: boolean
        aiScoreThreshold:
          type: integer
        pushEnabled:
          type: boolean
        pushScoreThreshold:
          type: integer
      required: [title]
    PhotoCenterItem:
      type: object
      properties:
        photoId:
          type: string
        filename:
          type: string
        thumbnailUrl:
          type: string
          description: 管控端 300px 缩略图 URL；当前过渡期可能仍是代理缩略图。
        derivativeStatus:
          type: string
          enum: [failed, pending, ready]
          description: 三规格派生图生成状态。
        sourceType:
          type: string
          enum: [local, feiniu]
        sourceAlbumId:
          type: string
        albumName:
          type: string
        aiScore:
          type: integer
          nullable: true
        aiScoreStatus:
          type: string
          enum: [pending, completed, failed]
        aiComment:
          type: string
        aiCommentStatus:
          type: string
          enum: [pending, completed, failed]
        aiTags:
          type: array
          items:
            type: string
        aiMemoryScore:
          type: integer
          nullable: true
          description: 情感记忆价值评分，来自统一 Vision AI。
        aiBeautyScore:
          type: integer
          nullable: true
          description: 构图与光影美感评分，来自统一 Vision AI。
        aiIsTrash:
          type: boolean
          description: 是否为废片。
        aiReason:
          type: string
          description: AI 简短理由。
        aiLayout:
          type: object
          properties:
            position:
              type: string
              enum: [top_left, top_right, bottom_left, bottom_right, center_safe]
            safeArea:
              type: object
              properties:
                x:
                  type: number
                y:
                  type: number
                w:
                  type: number
                h:
                  type: number
            textColor:
              type: string
              enum: ['#FFFFFF', '#000000']
            fontStyle:
              type: string
              enum: [serif, sans-serif, handwriting]
        removable:
          type: boolean
          description: true 表示可从播放相册移除；飞牛来源相册成员为 false。
      required: [photoId, thumbnailUrl, derivativeStatus, sourceType, aiScoreStatus, aiCommentStatus]
    UpdatePlaybackAlbumAiPolicyRequest:
      type: object
      properties:
        aiEnabled:
          type: boolean
        aiScoreThreshold:
          type: integer
        aiPriorityTags:
          type: array
          items:
            type: string
    PlaybackAlbumAiJob:
      type: object
      properties:
        jobId:
          type: string
        status:
          type: string
          enum: [queued, completed]
        requestedPhotoCount:
          type: integer
        importedSourcePhotoCount:
          type: integer
        generatedPhotoCount:
          type: integer
        skippedPhotoCount:
          type: integer
        startedAt:
          type: string
          format: date-time
        finishedAt:
          type: string
          format: date-time
      required:
        - jobId
        - status
        - requestedPhotoCount
        - importedSourcePhotoCount
        - generatedPhotoCount
        - skippedPhotoCount
    AiSettings:
      type: object
      properties:
        provider:
          type: string
        model:
          type: string
        baseUrl:
          type: string
        apiKeyConfigured:
          type: boolean
        scoringPrompt:
          type: string
        commentPrompt:
          type: string
        classificationPrompt:
          type: string
      required: [provider, model, baseUrl, apiKeyConfigured]
...
    UpdateAiSettingsRequest:
      type: object
      properties:
        provider:
          type: string
        model:
          type: string
        baseUrl:
          type: string
        apiKey:
          type: string
        scoringPrompt:
          type: string
        commentPrompt:
          type: string
        classificationPrompt:
          type: string
...
    PlaybackAlbumEnvelope:
      allOf:
        - $ref: '#/components/schemas/EnvelopeBase'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/PlaybackAlbum'
          required: [data]
    PlaybackAlbumListEnvelope:
      allOf:
        - $ref: '#/components/schemas/EnvelopeBase'
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/PlaybackAlbum'
          required: [data]
    PhotoCenterItemsEnvelope:
      allOf:
        - $ref: '#/components/schemas/EnvelopeBase'
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/PhotoCenterItem'
          required: [data]
    PhotoCenterItemEnvelope:
      allOf:
        - $ref: '#/components/schemas/EnvelopeBase'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/PhotoCenterItem'
          required: [data]
    PlaybackAlbumAiJobEnvelope:
      allOf:
        - $ref: '#/components/schemas/EnvelopeBase'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/PlaybackAlbumAiJob'
          required: [data]
    PlaybackAlbumAiPolicyUpdateEnvelope:
      allOf:
        - $ref: '#/components/schemas/EnvelopeBase'
        - type: object
          properties:
            data:
              type: object
              properties:
                album:
                  $ref: '#/components/schemas/PlaybackAlbum'
                aiJob:
                  $ref: '#/components/schemas/PlaybackAlbumAiJob'
              required: [album]
          required: [data]
    AiSettingsEnvelope:
      allOf:
        - $ref: '#/components/schemas/EnvelopeBase'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/AiSettings'
          required: [data]
...